+++ test/math_operations_service_test.go
@@ -9,9 +9,9 @@ import (
 
 func TestDoMath(t *testing.T) {
 
-	firstNumber := int64(10)
-	secondNumber := int64(5)
-	var result int64
+	firstNumber := float32(10)
+	secondNumber := float32(5)
+	var result float32
 
 	mathOperationsService := mathOperations.NewMathOperationsService()
